+++ src/engine/incremental/resources_state/mod.rs
@@ -4,7 +4,6 @@ mod fs;
 use crate::domain::Resources;
 use anyhow::Result;
 use serde::{Deserialize, Serialize};
-use std::path::Path;
 
 #[derive(Serialize, Deserialize, PartialEq)]
 pub struct ResourcesState {
@@ -13,16 +12,16 @@ pub struct ResourcesState {
 }
 
 impl ResourcesState {
-    pub fn current(resources: &Resources, project_dir: &Path) -> Result<Self> {
+    pub fn current(resources: &Resources) -> Result<Self> {
         Ok(Self {
-            fs: fs::ResourcesState::current(resources, project_dir)?,
-            cmd_stdout: cmd_stdout::ResourcesState::current(resources, project_dir)?,
+            fs: fs::ResourcesState::current(&resources.paths)?,
+            cmd_stdout: cmd_stdout::ResourcesState::current(&resources.cmds)?,
         })
     }
 
-    pub fn eq_current_state(&self, resources: &Resources, project_dir: &Path) -> Result<bool> {
+    pub fn eq_current_state(&self, resources: &Resources) -> Result<bool> {
         // TODO Parallelize this computation
-        Ok((&self.fs).eq_current_state(resources, project_dir)?
-            && (&self.cmd_stdout).eq_current_state(resources, project_dir))
+        Ok((&self.fs).eq_current_state(&resources.paths)?
+            && (&self.cmd_stdout).eq_current_state(&resources.cmds))
     }
 }