+++ tests/integ.rs
@@ -1,4 +1,5 @@
 use assert_cmd::prelude::*;
+use predicate::str::contains;
 use predicates::prelude::*;
 use std::ffi;
 use std::process::Command;
@@ -8,7 +9,7 @@ fn circular_dependency() {
     zinoma_command("circular_dependency", &["target_1"])
         .assert()
         .failure()
-        .stderr(predicate::str::contains("Circular dependency"));
+        .stderr(contains("Circular dependency"));
 }
 
 #[test]
@@ -16,7 +17,7 @@ fn imports() {
     zinoma_command("imports", &["target_2"])
         .assert()
         .success()
-        .stdout(predicate::str::contains("This is target 1"));
+        .stdout(contains("This is target 1"));
 }
 
 #[test]
@@ -24,7 +25,7 @@ fn named_project() {
     zinoma_command("named_project", &["target_3", "my_project::target_3"])
         .assert()
         .success()
-        .stdout(predicate::str::contains("This is target 1"));
+        .stdout(contains("This is target 1"));
 }
 
 #[test]
@@ -32,10 +33,8 @@ fn non_matching_imported_project_name() {
     zinoma_command("non_matching_imported_project_name", &["--clean"])
         .assert()
         .failure()
-        .stderr(predicate::str::contains(
-            "Failed to import incorrect_subproject_name",
-        ))
-        .stderr(predicate::str::contains(
+        .stderr(contains("Failed to import incorrect_subproject_name"))
+        .stderr(contains(
             "The project should be imported with name subproject_name",
         ));
 }
@@ -45,10 +44,8 @@ fn import_project_with_no_name() {
     zinoma_command("import_project_with_no_name", &["--clean"])
         .assert()
         .failure()
-        .stderr(predicate::str::contains("Failed to import noname"))
-        .stderr(predicate::str::contains(
-            "Project cannot be imported as it has no name",
-        ));
+        .stderr(contains("Failed to import noname"))
+        .stderr(contains("Project cannot be imported as it has no name"));
 }
 
 #[test]
@@ -56,7 +53,7 @@ fn invalid_project_name() {
     zinoma_command("invalid_project_name", &["--clean"])
         .assert()
         .failure()
-        .stderr(predicate::str::contains(":::: is not a valid project name"));
+        .stderr(contains(":::: is not a valid project name"));
 }
 
 #[test]
@@ -64,13 +61,13 @@ fn root_input_path() {
     zinoma_command("root_input_path", &["--clean", "print_source"])
         .assert()
         .success()
-        .stdout(predicate::str::contains("Content of my source file"));
+        .stdout(contains("Content of my source file"));
 
     zinoma_command("root_input_path", &["print_source"])
         .assert()
         .success()
-        .stdout(predicate::str::contains("Content of my source file").not())
-        .stderr(predicate::str::contains("Build skipped (Not Modified)"));
+        .stdout(contains("Content of my source file").not())
+        .stderr(contains("Build skipped (Not Modified)"));
 }
 
 #[test]
@@ -86,14 +83,14 @@ fn cmd_stdout_input() {
     zinoma_command(integ_test_dir_name, &["--clean", "changing", "stable"])
         .assert()
         .success()
-        .stderr(predicate::str::contains("changing - Build success"))
-        .stderr(predicate::str::contains("stable - Build success"));
+        .stderr(contains("changing - Build success"))
+        .stderr(contains("stable - Build success"));
 
     zinoma_command(integ_test_dir_name, &["changing", "stable"])
         .assert()
         .success()
-        .stderr(predicate::str::contains("changing - Build success"))
-        .stderr(predicate::str::contains("stable - Build skipped"));
+        .stderr(contains("changing - Build success"))
+        .stderr(contains("stable - Build skipped"));
 }
 
 #[test]
@@ -101,7 +98,7 @@ fn dependency_output_as_input() {
     zinoma_command("dependency_output_as_input", &["--clean", "print"])
         .assert()
         .success()
-        .stdout(predicate::str::contains("Intermediate build result"));
+        .stdout(contains("Intermediate build result"));
 }
 
 #[test]
@@ -112,7 +109,50 @@ fn circular_dependency_in_resources() {
     )
     .assert()
     .failure()
-    .stderr(predicate::str::contains("Circular dependency"));
+    .stderr(contains("Circular dependency"));
+}
+
+#[test]
+#[cfg(target_os = "macos")]
+fn incremental_multi_projects_build() {
+    zinoma_command(
+        "incremental_multi_projects_build",
+        &["--clean", "print_outputs"],
+    )
+    .assert()
+    .success()
+    .stderr(contains("print_outputs - Build success"));
+
+    zinoma_command("incremental_multi_projects_build", &["print_outputs"])
+        .assert()
+        .success()
+        .stderr(contains("print_outputs - Build skipped"));
+
+    zinoma_command(
+        "incremental_multi_projects_build",
+        &["--clean", "cmd_output::build"],
+    )
+    .assert()
+    .success()
+    .stderr(contains("cmd_output::build - Build success"));
+
+    zinoma_command("incremental_multi_projects_build", &["print_outputs"])
+        .assert()
+        .success()
+        .stderr(contains("print_outputs - Build success"));
+
+    zinoma_command(
+        "incremental_multi_projects_build",
+        &["--clean", "fs_output::build"],
+    )
+    .assert()
+    .success()
+    .stderr(contains("fs_output::build - Build success"));
+
+    zinoma_command("incremental_multi_projects_build", &["print_outputs"])
+        .assert()
+        .success()
+        .stderr(contains("print_outputs - Build success"));
 }
 
 fn zinoma_command<I, S>(integ_test_dir_name: &str, args: I) -> Command